pyproject.toml
[tool.poetry]
name = "sph"
version = "0.4.5"
description = ""
authors = ["François Poizat <francois.poizat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
setuptools = "^62.1.0"
GitPython = "^3.1.27"
colorama = "^0.4.4"
halo = "^0.0.31"
xdg = "^5.1.1"
PyGithub = "1.54"
PyYAML = "5.4.1"
cursor = "^1.3.5"
ratelimit = "^2.2.1"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
pylama = "^8.3.8"
pylint = "^2.13.8"
jedi = "^0.18.1"
flake8 = "^4.0.1"
bandit = "^1.7.4"
kalong = "^0.3.1"

[tool.poetry.scripts]
sph = "sph.sph:be_helpful"
workflow = "sph.workflow:be_helpful"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
